+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Any regex group matches are now passed in the [Arguments] parameter to the function/command.
 
+### Fixed
+- Help syntax now reflects the command name (as known in PoshBot) instead of PowerShell function name.
+
 ### Changed
 - Command usage help is now displayed differently according to the command's trigger type.
   For [Command] trigger types, the command/function's parameters are show, for [Regex] trigger